fix: added test utility functions and cleanup
